Created a asyncTask which steps through a loop of all receipts in the folder, first it inflates the receipt activity layout, next it makes a php call to get all related information for the receipt, next it populates the layout with the new information, finally it takes a screenshot of the layout and saves it to the users gallery. But has a problem of cutting of some words throughout the receipt.
Created a class with a function similar to the in-build "MediaStore.insertImage" This function allows me to push all receipts from a folder into one album related to that folder, in the users gallery.
Also now all single exported receipts go into a album called "All Receipts"

Added check to see if permission is asked for exporting before creating the image.